Add ability to load extra data server side
Added the enhancedLoadFunction property to the options that can
be passed in to this middleware.  The enhancedLoadFunction can
be used to load extra data or do whatever manipulations are
needed on whatever object is added to the request object.

An example use case would be storing a minimum of data points on
the JSON Web Token (JWT), for example nothing more than an
identifier into some cache mechanism server-side (e.g. memory,
memcache, redis) that, when the JWT is verified, can then be
extracted from the decoded JWT data and used to look up
additional values from whatever server-side persistence store
might be used.

The new callback is a "standard" node callback that takes in the
decoded JWT object with its properties and a callback, and can
then do whatever it wants with the object and its properties, and
then calls the given callback with an error object (null or falsey
if no error) and the decoded object with whatever changes the
callback may or may not have made to the decoded object.  Example
errors might be that the session has been expired (early) server-
side, or is not found.